requirements.txt
matplotlib
numpy
pandas
tqdm
bioframe
cooler
cooltools
pybbi
